Rename AddDebuggingSession component
