+++ frontend/src/components/AddDebuggingSessionModal.vue
