+++ frontend/src/views/main/project/Debugger.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import { api } from '@/api';
 import { InspectionDTO } from "@/generated-sources";
-import AddDebuggingSession from '@/components/AddDebuggingSession.vue';
+import AddDebuggingSessionModal from '@/components/AddDebuggingSessionModal.vue';
 import InspectorWrapper from './InspectorWrapper.vue';
 import InlineEditText from '@/components/InlineEditText.vue';
 import DeleteDebuggingSession from '@/components/DeleteDebuggingSession.vue';
@@ -117,7 +117,7 @@ onActivated(() => loadDebuggingSessions());
             <v-btn v-if="!displayComponents" @click="showPastSessions" class="mr-4 pa-2 text--secondary">
               <v-icon left>history</v-icon>Past sessions
             </v-btn>
-            <AddDebuggingSession v-bind:project-id="projectId" v-on:createDebuggingSession="createDebuggingSession"></AddDebuggingSession>
+            <AddDebuggingSessionModal v-bind:project-id="projectId" v-on:createDebuggingSession="createDebuggingSession"></AddDebuggingSessionModal>
           </div>
         </v-col>
       </v-row>
@@ -163,7 +163,7 @@ onActivated(() => loadDebuggingSessions());
       <v-alert class="text-center">
         <p class="create-session-message headline blue-grey--text">You haven't created any debugging session for this project. <br>Please create your first session to start debugging your model.</p>
       </v-alert>
-      <AddDebuggingSession v-bind:project-id="projectId" v-on:createDebuggingSession="createDebuggingSession"></AddDebuggingSession>
+      <AddDebuggingSessionModal v-bind:project-id="projectId" v-on:createDebuggingSession="createDebuggingSession"></AddDebuggingSessionModal>
       <div class="d-flex justify-center mb-6">
         <img src="@/assets/logo_debugger.png" class="debugger-logo" title="Debugger tab logo" alt="A turtle using a magnifying glass">
       </div>
